requirements.txt
Django==4.2
django-ajax-selects==2.2.0
django-bootstrap4==23.1
django-tables2==2.5.3
django-bootstrap-datepicker-plus==5.0.3